requirements.txt
aiohttp~=3.6.2
aioredis~=1.3.1
gino~=1.0.1
celery~=4.4.6
Pillow~=7.1.2
pydantic~=1.5.1
redis==3.5.3
pytest~=5.4.3
pytest-aiohttp==0.3.0

